Lint

+++ init.py
@@ -44,9 +44,9 @@ units.FTE = FTE
 
 time_horizon = 20   # years
 
-v_zeros = np.zeros(time_horizon + 1, dtype=np.float64)
-v_ones = np.ones(time_horizon + 1, dtype=np.float64)
-v_after_invest = np.ones(time_horizon + 1, dtype=np.float64)
+v_zeros = np.zeros(time_horizon + 1)
+v_ones = np.ones(time_horizon + 1)
+v_after_invest = np.ones(time_horizon + 1)
 v_after_invest[0] = 0
